Revolvers use six shots. Not seven.
I'm aware some exceptions exist, but according to the code, we're using
a The Lumoco Arms HE Colt, which we can only assume is derived from an
existing Colt revolver.

If we say "No it's special snowflake sci-fi revolver with SEVEN shots"
then why does the Mateba use seven shots too? The Mateba doesn't use
seven shots.

If spacemen revolvers have seven shots, why doesn't the detective's
revolver use seven shots?

Either the default revolver uses 7 shots and is a fancy Russian revolver
and none of the others are, or all revolvers are 6 shot. It seems silly
to have it both ways.