Reverts a hyperzine change
Hyperzine makes you fast. Due to some change a while back, it now makes you 3X as fast as it used to. This results in players moving _at the speed of fucking light_ instead of being a chemical that was helpful but not insanely over powered.

Hyperzine is easy to make (Literally 3 clicks in the chem dispencer), can be made into 14.88U pills to prevent ODing, and lasts bloody ages to fully metabolize.

This PR makes it so hyperzine makes you a _little_ faster, but not speeds that puts sonic to shame,