Coded the flower crowns.
Added wood circlet, craftable with 1 plank of wood.

If poppy, sunflower, or lavender seeds are used on the circlet, it makes a flower crown appropriate for the flower type. From what me and Cerebulon see, it's not possible to refer to flowers directly. Feel free to tweak the coding if you'd like.