Changes 5.56 to 5.45
Because the STS is clearly a goddamn Space AK and I cringe every time I
see 5.56 on its ammo type.

This change serves to do literally nothing except make gun enthusiasts
like me not cry inside, even though it's a fictional space gun that
nobody cares about but DAMN IT I CAN USE THAT LOGIC TOO, SO IT USES
5.45mm MUNITIONS NOW. :U

Also fixes a few oversights like how the assault rifle was still labeled
5.56 (or now 5.45) in the traitor menu, or how the ceremonial rifle said
it used the wrong ammo as well.

Oh yea, and the small magazine for the 5.45 gun actually looks like it's
the same mag for the gun. This has the added bonus of freeing up that
SVD mag for later use, if we ever want it.