Added emergency bolt override to airlock code
Airlocks now have an emergency_override var that gives the option to manually alter the bolt state when the airlock's unpowered by cranking the override with a wrench.

By default only external airlocks have this enabled so that people don't have to break a window to bypass inaccessible external airlocks anymore.