Another possible crash bug breakthrough.
So we had an incident that brought definite results pointing to a certain circumstances with an observer successfully managing to crash everyone with an impressively high rate of success. This observer ghost was a teshari wearing a straightjacket. Some comparisons to other less definite cases were made, pointing towards characters with species specific tails arriving in gear that hides tails. Some code diving later I came to a conclusion that the code part responsible for generating cache for tail icons was locked behind the suit coverage check, suggesting that arriving onscreen for the first time while wearing one of those suits would prevent the tail part of the cache from generating at all.