Fixes missing hollow point ammo for 9mm
9mm didn't have HP ammo for some reason, which is dumb, considering HP
ammo makes 9mm not useless while not being overpowered as hell.

All I did was add in the missing projectiles and relevant casings. How
they get implimented if at all is up to another PR.

Also made some notes for later because apparently Polaris is in a
feature freeze so I am not pushing new features, just fixing a missing
one, but I need to remind myself to do this later or remind somebody
else if I forget.