Add scheduled callbacks
And include a callback for setting vars.

You'd use it like `schedule_callback_in(5 SECONDS, VARSET_CALLBACK(src, myvarname, FALSE))` if you wanted to set `myvarname` to `FALSE` in 5 seconds.

This was originally going to be a bugfix to you but I guess I totally forgot that I didn't actually send this to you originally.