Kraut Space Magic™ and other gun changes
#It's that time again, when Ace mucks with all the guns

This is probably going to need a changelog but I forget how to do that.
Somebody guide me through that later.

Also it's 12:30AM and I'm facedesking. I'll finish a proper description
for this PR when I am awake and functional again. For now...

- 5.45 is gone. Long live .223!
- 7.62 is now .308 as well. I'll explain why I use this format later.
- SVD now uses its own ammo, which uses hitscan, so it's not useless as
a sniper.
- Fixed a whole lot of shit. Will elaborate later.
- Made a bunch of quality of life improvements. Will elaborate later.
- Added support for caseless ammunition, and a G11 knockoff to showcase
it. G11 was done by somebody on Discord named Ryo who happens to love
the G11. I actually think it's ugly as shit and I hate everything about
it, but it was the perfect candidate to introduce caseless ammo to the
game. Also KRAUT SPACE MAGIC™ meme.
- asjklgfljakdhjsk
- https://www.youtube.com/watch?v=KufiqCh3Qd4
- i need sleep