Replaces 'City Supervisor' with 'City Clerk'
This update replaces City Supervisor with City Clerk, as the Mayor is meant to be the supervisor and manager of the city (As per Cassie).

This update also changes the Judge's roundstart 'your supervisor is' text to be nothing, instead of the 'City Clerk' - as the Judge is independent of the city's structure and is the head of the local judiciary.