Updates Janitor and Judge
Replaces 'Janitor' with 'Sanitation Technician and adds the following titles:
- Recycling Technician
- Sanitation Engineer.

These are more city-oriented titles.

Changes the Judge's supervisor from the City Clerk, to 'government officials and the President' like the mayor, as the Judge does not answer to city officials as they are the head of the local judiciary.