Copy character slots
Adds a button to the setup menu that can be used to copy a character over another slot, allowing for easier duplicating, rearranging, and wiping of character slots. After you copy a character to another slot, you must then save to overwrite what's already on that slot - or don't, if you've changed your mind.