Infinite Frame Fixy
On the tin: fixes production of infinite frames by porting https://github.com/Citadel-Station-13/Citadel-Station-13-RP/pull/855. This also eliminates the given runtimes of being unable to read a board type if it wasn't given for the machine object. Additionally allows machines related to Hydroponics to be unwrenched while refusing to create frames when crowbar'd, returning a whack instead. Tested and it works.