Nerfs Simple Doors
Simple doors would take the integrity of the material (150 for steel AND resin), divide it by 10, and use that hardness as health. Damage taken by the door would have its force divided by 100. This meant that in order to do 15 damage to a steel door with a baton, you'd need to hit the door 100 times. Now you only need to hit the door 10 times.

Also makes it so bullets/lasers work against the doors

Summation: Resin doors take 10 baton hits to break. (Resin walls take 14.)