alters tables.dm
Still issues: If you harm intent someone against a table while holding a grab, it doesn't slam or possibly break the table. This is functionality that was removed, but I assume there was a reason, so I'm not re-adding it.

Also none of these issues are my fault. BUT HERE'S A FIX ANYWAY.