Orange Is The New Orange Part 2
Refactors prison jumpsuit slightly to not be a child of the orange jumpsuit, which in turn gets its own sprite. Left/right hand uniforms don't seem to be applying in-game when carrying the suits but that's not too bad of an issue. Uniform on the mob shows properly and works with sleeves rolled and uniform rolled down.