Cleans up cryobags
Cryobags are two different things: One of them is the object and the other is a subtype of closets. When I cleaned up closet code, I failed to tackle cryobags, because they were in a completely different location.

This is why the file hierarchy needs to be arranged like the object hierarchy 99.99% of the time.

I can honestly say cryobag closet code was a lot cleaner than either closet or crate code.