Hotfixes pits
It's not a clean fix. It does some things I don't like - not giving feedback when you pull things out of the pit, making them appear under your feet instead of ontop of the pit itself, no mechanic for inserting tools like shovels in - but for now, it doesn't do buggy stuff and actually buries things.

So it functions.