Gundraw commentary
Closed the old PR because merge conflicts with polaris.dme and it got super ugly.

Commented out code for setting a weapondraw delay based on how heavy the weapon is. Personal preference is to remove instantaneous firing and switch back to slow-moving pellets and not do this ; ditto with throwing.

Would like a declaration as to which direction is being gone before implementing.

Other than that, some commentary on how certain things could be cleaned up that I may or may not get to.