Removes Extended
Yes. It is actually happening. This gamemode is going to the dogs, ladies and gentlemen.

...

Actually we're replacing it with the Canon gamemode, which is basically the same as extended but your persistent data saves, and you can do a bit of conflict as long as it doesn't go outside of our RP guidelines. Freeform RP - what extended was supposed to be.

Actually removing extended because of the connotations of extended implying that you are supposed to sit around and chair RP, it doesn't have to be that way. Have fun.