Makes Bank Accounts Fully Persistent
The idea of this is so that bank accounts can be accessed when people aren't around by the game's code.

Which means this allows devs to make code where people recieve or lose money while offline. This might mean we'll have to put in some temporary transisition code, but it should be fun and good once we get it working.

Since this is gonna be a biggie, might also look into fingerprint scanners for ATMs.