Stops people from stripping SSD players (Unless Antag/Police)
Port from Paradise. This does what it says on the tin. Admins will have a button to toggle this setting off/on, and it's also a configuration option that can be disabled in general if needed.

Antagonists and police can strip SSDs, within reason, but those roles are jobbannable if they are abused so we'll see how it goes.