Prettifies closet contents list with (now added) countable english_list
Contents of closets (as viewed by ghosts) are now less ugly and easier to read.

- Adds reusable option to have english_lists count the contents and
display them (optionally with icons and specific determiners).
- Due to backwards compatibility these options are NOT available for
non-countable lists; this was originally planned, but it's not easily
possible unless all usage of english_list is revised.
- Moves the ghost examine code to proc/examine where it belongs.
- Ghosts now also see empty/full messages like characters standing
close do.
- The contents list for closets now shows number of items (if multiple)
and their icons (wherever possible).
- The list no longer shows unnecessary determiners ("the" before every
single item).
- Old usage of english_list still uses the old code. New code could
change order of the strings where it matters (in weather reports or
camera network assembly code, for example), so it cannot be used.