Fix Runtime in loadout.dm,277: bad index
Runtime happens when someone has a saved preferences loadout containing an item that currently is expected to have gear_tweaks, but at the time that the savefile was last saved, did *not* have gear tweaks.  Thus the saved prefs have null metadata.
Also it makes no sense to check the length of a list *inside* a for loop iterating over that list..