Nerfs Chameleon Kit Boxes
Because of the way chameleon kits start with items, the boxes start with a ton of storage slots to hold all the items. Sometimes people grab the boxes, dump out the original contents and stuff a ton of stuff into them as a result.

This refactors it so that all the chameleon stuff starts in the backpack instead, so chameleon kit boxes themselves are no longer superboxes