Ports Inactivity SS from Bay
Remember that server-destroying timer bug that VORE keeps getting?

Turns out, it is related to the inactivity subsystem, that, due to having SS_BACKGROUND and SS_NO_TICK_CHECK flags, causes the master controller to break in a horrible way. One of the consequences of the master controller breaking is that the timer subsystem stops running.

I am relieved that this monster of a bug has finally been slain. Hopefully.