Helmet Camera Verb No Longer Shows Without Helmet Camera
Changes the toggle helmet camera verb to a proc

Adds a check in the initialization of spacesuit helmets to see if it has a camera network.

Removes said check from the proc itself.

Basically this means you won't get the toggle-camera verb if you don't have one on your helmet.