Allows backpacks to be carried on most hardsuits - Backpacks can be carried on suit storage slot of those hardsuits that occupy back slot (breacher excluded). - Now, i understand this may be a somewhat controversial change. Feel free to close this PR if you believe it does not fit. Just opening this as i've received this suggestion from another player, and i in fact tried proposing the same thing back at Bay when hardsuits as we know them were added. - Reformatted some of the allowed lists while i was at it. Makes future modifications somewhat less likely to hit merge conflicts with each entry on its own line.
My balance perspective (long rant):
Hardsuits, as-is, fill very specific roles. In general, unless you absolutely need one of the modules or some specific function (for example, 100% RAD resist combined with EVA capability of CE suit) you don't want to use them, as they require you to trade-in your back slot (which is what this PR aims to change), therefore significantly reducing carrying capacity. Of course, this has its workarounds - usually dragging crates around, or carrying your backpack in hand. This provides minimal actual restriction (you can still carry tons of stuff), but acts as quite major inconvenience and annoyance for the player (you have to juggle with items, keep placing stuff on the ground, etc - generally just clicking 10x more). In total, the only real effect is that it slows you down a bit due to all the clicking required.

Therefore i believe this change is not as massive buff as it might seem in first place. Sure, it lets you carry more things on yourself, but you could also do that easily with a crate or by holding backpack in hand. It just makes doing so more convenient and comfortable to the player.