Streamline (Facial) Hair & Marking Character Customization
Adds the ability to shift to the previous or next facial hair style. The buttons are behind the style because this way you don't have to move your mouse when the style changes.

If you click forward while on the last style, it'll shift to the first one.
If you click backward while on the first style, it'll shift to the last one.

For markings, you can now pick a marking and inject it infront of another so long as it isn't the one you picked, or the one directly behind in the list order.