Add the Aurora Caelus Event!
Adds the Aurora Caelus, a rare event that occurs on the station space z-levels~!

During this event a .ogg will play to all clients, and you will see the outside appear almost like a storm - consider it a fun little break and a rare event for a moment of calm.

The event announces as soon as it fires, and assuming time dilation isn't horrible, starts 120 seconds after (60 "ticks" of the event subsystem being fired by the MC, which happens every 2 IRL seconds, affected by time dilation).

Current event weight (frequency) can be tweaked according to admin liking, and it's currently a MUNDANE level event. This can be changed to MODERATE or otherwise as determined by admins, or given a higher weight - during testing, stuff like PDA spam or news or etc very often went over it and it took giving it a lot of weight for it to fire - but it is supposed to be a "rare" cosmic thing, not something that happens every shift.