Rag Coding Fix
A quick one file fix for a rag coding issue related to face covering helmets not providing protection to smothering should rags be filled with chloral hydrate.  Basic face masks will not protect against the intent, but items like welding masks and space suit helmets will.

This is a fix for an issue brought up on a server downstream, pulled here to hopefully prevent any future conflicts arising should this file be modified in the future.