Fixes null grab runtime breaking mob movement
Adds a sanity check to prevent the client move proc from runtiming between setting and clearing the "busy time" between movements. In other words, the proc sets "moving" to 1 to prevent other movement before the current one is finished and having a null in the grablist will runtime the proc before it can set "moving" back to 0, effectively leaving the client locked in that "busy time" for good.