5 POI's
As the tittle, adds 5 additional POI's intended for the plains.
Nothing valuable or useful should be found on any of these, maybe a bit of story and curiosity for anyone that wants to explore these.
Or better... maybe explore the remains and restore them to its former glory and take ownership to any abandoned home.