Improves Portal Documentation for New Mappers unfamiliar with DIR
This isn't really needed, but more of a safety thing. I could not puzzle out the "opposite directions" until I checked the code and realized the portals forcemove you in the DIR it's facing. Therefore, if the portal was facing north, but the north is outside the map, it'd runtime.

This documentation should help new mappers, even if it is a bit 'dumbed down'.