Adds Engine Randomization on Mapload!
Engines will now be randomly placed on mapload, maintainers will need to update the server config to enable the system.

Add this to Config:

\# Control which submaps are loaded for the Dynamic Engine system
ENGINE_MAP Supermatter Engine,Edison's Bane

This will load either Supermatter or Tesla based on randomization. If you'd prefer to just stick with SM, remove "Edison's Bane" from the config file and leave in Supermatter Engine.

The way this is done DOES require the engine load system to be used, but it should pave the way for more modular engines in the future!

This also allows older maps to be loaded with hard-coded engines, if you comment out the line in config to disable it, it will skip engine loading.

If there's any issues, refer to the relevant submaps to tweak 'em. <3

TL;DR - Supermatter is now changed to on-mapload, Tesla added as option for mapload engine start. Remove the "Edison's Bane" from config and put the entire line into your config to make the engine work now. :blep:

Please merge this quickly, it will produce conflicts.