rad suit sprite change + space suit helmet sprite
changes the sprite for the radiation suit the change of the hood from the humanoid's face shield to the tehari's full on dome seems odd to me. On top of that the side sprites for the suit itself are oddly inconsistent from the front/back. I also made a sprite fro the missing tesh space helmet some time ago.