Halves Item Slowdown Effects
Most items in the codebase have a slowdown multiplier of one, when slowdown is handled species-side in multiples of 0.5. This makes the effect of species slowdown almost non-existant when item slowdown is involved, and makes species that have item slowdown reduction (such as Unathi) unnaturally fast.

Do note that I also went through and removed odd slowdown numbers as well, as, due to the way slowdown is tied to tickrate, it should be done in intervals of 0.5.

I tried to view things on a case by case scenario but there will probably be changes that will need to be double checked.

Some items will be handled differently, I will list the exceptions below.
Boots of Speed: Kept at -1, it is boots of speed after all.
Chemistry Bag: Increased to 3. This should not really be toted around outside of chemistry.
Hybrid Tools: They already have a lot of other disadvantages, and are kind of hard to get, and 0.1 is a weird number for slowdown.
Site Manager's armor: Already 1.5
Unathi Breacher modules: These things are overpowered and Unathi already have a racial trait to virtually remove slowdown.
Voidsuits: Already done in intervals of 0.5, no change.
RHR Accellerator: This thing probably has appropriate slowdown values for what it is.
Utility suits: Already done in intervals of 0.5
Clothing/suits/armor: Already done in intervals of 0.5