Added plate carriers
Found out the plate carriers needed sprites to call on. I did not change the colors of each vest, I just copy pasted the 'bulletproof_new' vest and then renamed each of the copies for the different colored Plate Carriers, all of them are black but I don't know how to fix the colors easily without doing entire resprite which is not in my skill set.

If someone later chooses to fix them the base sprite is there.