Breathing Sounds While in Space
Upstream port of https://github.com/CHOMPStation2/CHOMPStation2/pull/3831

Plays a breathing sound akin to Bay's breathing sound while you're in space.

Major difference (codewise) is the ability to set an inhale + exhale sound, and it is currently designed SOLELY to be heard while in space.
Keep inhale + exhale combined to 7 seconds or less if you add custom sounds.
The code DOES support custom breathing per-species, if you see fit to add that for w/e reason.

Any internals playing breathing sounds later on may be looked at, but for now, this gets breathing working.

Does not come with a pref currently, as it's quiet - and intended to be an audible indicator of if you're breathing or not.
(IE if you don't hear it, be concerned.)