Antag Uplink Medical Adjustments
Major reduction on the cost of all medical items. Reason being traitors tend not to buy medical items since they're so expensive but they also don't have access to the same kind of medical care as the crew.

Also added the basic medical kits to the uplink. Costs will be adjusted if need be in the future.