add atom/Topic & atom/CreateAtomTopic + examine
I was originally going to do a suggestion but ended up writing an
alternative since it would've been a very big suggestion.

atom/Topic and atom/CreateAtomTopic are provided as a pair with basic
source verification built in. *This verification is not currently
applied for subtypes of atom because their own topic links do not pass
the appropriate reference.*

adds the AtomTopicExamine behavior & related code in atom examine.
Examined atoms that do not display all of their available information
in the chat box will indicate that they have more information available
and provide a link to switch to the examine tab of the stat panel. This
is applied for both preferences where the examine tab is not opened
automatically.

fixes a couple of display issues with examine behavior, including a
blank line showing when usage information should be included but none
was available.

closes #8808