Fixes a looping sound bugs, adds additional options
Previously, the pref check on `direct` loops would always fail if pref_check was not enabled, meaning that direct sounds never worked. This is fixed.

Modifies how _direct override works. Previously, it was simply doing direct = _direct, which meant that any and all sounds with direct defined would immediately be flipped back to false UNLESS you specified a specific argument.

Now, it is an +override+, labeled "disable_direct". If set true, this disables direct and forces it to use local playback.

Exclusive mode is added. Enable this on your looping sound to +only+ start one of this soundloop from your object at a time, rather than allowing multiple start() calls to fire and trigger new loops.

Skip Start Sound is an argument to start - this allows you to skip the startup and go straight to the mid.
Skip Stop Sound is similar to the above, but it skips the stop/end sound, and simply ends the loop.