adds novelty wet floor sign near loose stairs
up for negotiation if this gives it away too much.

my gut says it's largely novelty and a net gain. it'll make dying to the trap funnier in the worst case, and in the best case the clever will notice the unique text and look around for other things with unique text