Cleans up robot_upgrades.dm, nerfs VTEC module
Cleans up the formatting and style of robot_upgrades.dm.

Reduces the VTEC module's movement delay reduction from 1 to 0.5 as it's currently way too fast. This value is made a simple define so it can easily be changed in future.

Converts a lot of magic booleans to TRUE/FALSE.

Inverts the boolean logic of the /obj/item/borg/upgrade/proc/action base so it's consistent with all derived types.