Renames/re-paths HoP hat, adjusts locker contents
Re-paths the head of personnel's cap from /obj/item/clothing/head/caphat/hop to /obj/item/clothing/head/hop.

Lowers the amount of spare ID boxes in the head of personnel's locker from 2 to 1, since the additional is just mostly clutter.

Renames the "crew resource's hat" to the "head of personnel's cap" for consistency with the rest of the HoP equipment.