Updates head of security office layout
Moves the disposal unit from the east to the west corner of the room.

Moves the filing cabinet to the tile previously occupied by the disposal unit.

Adds a table, with the megaphone, universal recorder and cassette tape to the spot previously occupied by the filing cabinet.

Removes the megaphone, universal recorder and cassette tape from the head of security's locker.